pyproject.toml
[tool.poetry]
name = "video_streaming_django"
version = "0.1.0"
description = "django app"
authors = ["Nikita Zaretski <xal9wIII4ik@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
django-cors-headers = "3.11.0"
djangorestframework = "3.13.1"
psycopg2-binary = "2.9.2"
boto3 = "1.21.8"
fleep = "1.0.1"
djangorestframework-simplejwt = "^5.1.0"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"
mypy = "^0.941"
flake8 = "^4.0.1"
ipython = "^8.1.1"
moto = "^3.1.1"
mock = "^4.0.3"
types-mock = "^4.0.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
